+++ subpackages/cli/source/odood/cli/commands/database.d
@@ -13,9 +13,9 @@ private import commandr: Argument, Option, Flag, ProgramArgs;
 
 private import odood.cli.core: OdoodCommand, exitWithCode, OdoodCLIException;
 private import odood.lib.project: Project;
-private import odood.lib.odoo.lodoo: BackupFormat;
 private import odood.lib.odoo.test: generateTestDbName;
 private import odood.utils.odoo.serie: OdooSerie;
+private import odood.utils.odoo.db: BackupFormat;
 private import odood.utils: generateRandomString;
 private import odood.utils.addons.addon: OdooAddon;