+++ subpackages/lib/source/odood/lib/package.d
@@ -1,6 +1,6 @@
 module odood.lib;
 
-public immutable string _version = "0.0.15";
+public immutable string _version = "0.0.16";
 
 public import odood.lib.project;